* 🎉 chore(github): add README.md file * 🎉 chore(gitignore): add .gitignore file * 🎉 chore(docker): add .dockerignore file The README.md file has been added to the .github directory. The .gitignore file has been added to the root directory. The .dockerignore file has been added to the Docker directory. The .dockerignore file ignores all .git directories.
* 🐳 chore(Dockerfile): add Dockerfile for Madrestam
This commit adds a Dockerfile for Madrestam, which is a container for playing Ankama Games. The Dockerfile is based on jlesage/baseimage-gui:debian-11 and installs the Ankama Launcher and its dependencies. The Dockerfile also sets up sudoers and copies over some additional files.

* 🎉 feat(Docker): add setup script for app user and startapp script
* 🚀 chore(Docker): add build and run scripts
The `99-setup-app-user.sh` script sets the ownership of the `/akl` directory to the `app` user. The `startapp.sh` script executes the `zaap` binary located in `/akl/squashfs-root/`. The `build.sh` script builds the Docker image with the tag `akrista/madrestam`. The `run_example.sh` script runs the `akrista/madrestam` Docker image with the `--shm-size`, `--network`, and `--privileged` options.