* 🐳 chore(Dockerfile): add environment variables and install Steam * ✨ feat(Dockerfile): add environment variables for XDG directories and install Steam The environment variables APP_NAME and DARK_MODE are added to the Dockerfile. Additionally, environment variables for XDG directories are added to the Dockerfile to allow for better compatibility with the host system. Steam is installed by downloading the latest version from the Steam repository and installing it using make.
* 📝 chore(Dockerfile): update metadata labels
The metadata labels have been updated to reflect the correct name and description of the Docker container. The name has been changed to "MadreSteam" and the description has been updated to "Docker container for playing Steam".

* 🐛 fix(99-setup-app-user.sh): add newline at end of file
* 🎨 chore(main-window-selection.xml): change window title from Lindo to MadreSteam
* 🎨 chore(startapp.sh): refactor startapp.sh to use environment variables and launch steam instead of Lindo
* 🎨 chore(build.sh): refactor build.sh to use correct image name
* 🎨 chore(run_example.sh): refactor run_example.sh to use correct image name and mount /steam volume
The changes in these files are mostly refactoring and renaming. The changes in startapp.sh are significant as it changes the application that is launched from Lindo to Steam. The changes in run_example.sh are also significant as it changes the image name and mounts the /steam volume.